[update]HPをfloatに変更

+++ Assets/_Miyamoto/Scripts/EnemyData.cs
@@ -1,3 +1,4 @@
+using UnityEditor.ShaderGraph.Internal;
 using UnityEngine;
 /// <summary>
 /// “G‚Ì‰Šú’l‚ğŠÇ—‚·‚éƒf[ƒ^ƒNƒ‰ƒX
@@ -8,20 +9,20 @@ public class EnemyData : ScriptableObject
     [Header("“G‚Ì‰Šú’l")]
     [SerializeField] private int _enemyHp;
     [SerializeField] private float _enemyMoveSpeed;
-    [SerializeField] private int _enemyFov;
-    [SerializeField] private int _enemyPower;
+    [SerializeField] private float _enemyFov;
+    [SerializeField] private float _enemyPower;
     [SerializeField] private float _enemyAttackRange;
     [Header("“G‘ÎŠÖŒW")]
     [SerializeField] private EnemyState _enemyState;
 
     /// <summary>“G‚ÌHP‰Šú’l</summary>
-    public int EnemyHpData => _enemyHp;
+    public float EnemyHpData => _enemyHp;
     /// <summary>“G‚ÌˆÚ“®‘¬“x‚Ì‰Šú’l</summary>
     public float EnemyMoveSpeedData => _enemyMoveSpeed;
     /// <summary>“G‚Ì‹–ì‚Ì‰Šú’l</summary>
-    public int EnemyFoVData => _enemyFov;
+    public float EnemyFoVData => _enemyFov;
     /// <summary>“G‚ÌUŒ‚—Í‚Ì‰Šú’l</summary>
-    public int EnemyPowerData => _enemyPower;
+    public float EnemyPowerData => _enemyPower;
     /// <summary>“G‚ÌUŒ‚”ÍˆÍ‚Ì‰Šú’l</summary>
     public float EnemyAttackRangeData => _enemyAttackRange;
     /// <summary>“G‘ÎŠÖŒW</summary>
